// block-performance-1.0.0/src/library/packager/dockerfile-make/type.ts
export interface Input {
    name: string;
    version: string;
    with_customize_dir: boolean;
    custom_build_steps: { library_name: string; function_name: string; function_input: { [key: string]: any } }[];
}

export interface OutputOk {
    text: string;
}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}